refactor(FE): `useStartBattle` hook
